Rename json-test to json-encoding-test
