+++ test/index.js
@@ -32,7 +32,7 @@ require('./get-put-del-test')
 require('./idempotent-test')(test, testCommon)
 require('./init-test')
 require('./custom-encoding-test')(test, testCommon)
-require('./json-test')(test, testCommon)
+require('./json-encoding-test')(test, testCommon)
 require('./key-value-streams-test')
 require('./maybe-error-test')(test, testCommon)
 require('./no-encoding-test')
