+++ test/json-encoding-test.js
